# docs/swagger.yaml
basePath: /
definitions:
...
  internal_modules_auth.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/internal_modules_auth.User'
    type: object
  internal_modules_auth.AuthSuccessResponse:
    properties:
      data:
        $ref: '#/definitions/internal_modules_auth.AuthResponse'
      errors:
        type: object
      message:
        example: Logged in successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  internal_modules_auth.ErrorResponse:
    properties:
      data:
        type: object
      errors:
        type: object
      message:
        example: Validation failed
        type: string
      success:
        example: false
        type: boolean
    type: object
  internal_modules_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_auth.RegisterRequest:
    properties:
      email:
        maxLength: 190
        type: string
      name:
        maxLength: 120
        minLength: 2
        type: string
      password:
        maxLength: 72
        minLength: 8
        type: string
      username:
        maxLength: 80
        minLength: 3
        type: string
    required:
    - email
    - name
    - password
    - username
    type: object
  internal_modules_auth.User:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      role:
        type: string
      status:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: API documentation for EDNIC backend.
  title: EDNIC Backend API
  version: 0.1.0
paths:
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT access token.
      parameters:
      - description: Login request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
      summary: Login user
      tags:
      - Auth
...
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with role user.
      parameters:
      - description: Register request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_auth.AuthSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_auth.ErrorResponse'
      summary: Register user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"